Lógica de inicio de test activos se envía a fragments
