+++ app/src/main/java/cl/niclabs/adkintunmobile/views/activemeasurements/viewfragments/settingsfragments/SpeedTestSettingsFragment.java
@@ -1,7 +1,9 @@
 package cl.niclabs.adkintunmobile.views.activemeasurements.viewfragments.settingsfragments;
 
 import android.os.Bundle;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
 import android.support.v7.preference.Preference;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,11 +12,15 @@ import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.Toast;
 
+import java.net.HttpURLConnection;
+
 import cl.niclabs.adkintunmobile.R;
 import cl.niclabs.adkintunmobile.data.persistent.activemeasurement.SpeedTestReport;
 import cl.niclabs.adkintunmobile.utils.activemeasurements.speedtest.ActiveServersDialog;
 import cl.niclabs.adkintunmobile.utils.activemeasurements.speedtest.ActiveServersTask;
+import cl.niclabs.adkintunmobile.utils.activemeasurements.speedtest.CheckServerTask;
 import cl.niclabs.adkintunmobile.views.activemeasurements.ActiveMeasurementsActivity;
+import cl.niclabs.adkintunmobile.views.activemeasurements.SpeedTestDialog;
 
 public class SpeedTestSettingsFragment extends ActiveMeasurementsSettingsFragment{
 
@@ -26,12 +32,7 @@ public class SpeedTestSettingsFragment extends ActiveMeasurementsSettingsFragmen
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         this.context = getActivity();
         LinearLayout view = (LinearLayout) super.onCreateView(inflater, container, savedInstanceState);
-        Button startButton = addStartButton(view, context);
-        startButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                ((ActiveMeasurementsActivity) getActivity()).onSpeedTestClick();
-            }
-        });
+        addStartButton(view, context);
 
         return view;
     }
@@ -64,5 +65,54 @@ public class SpeedTestSettingsFragment extends ActiveMeasurementsSettingsFragmen
             Toast.makeText(context, getString(R.string.settings_active_measurements_reports_delete_toast), Toast.LENGTH_SHORT).show();
         }
         return super.onPreferenceTreeClick(preference);
+    }
+
+    @Override
+    public void onStartTestClick(){
+        if (!ActiveMeasurementsActivity.enabledButtons)
+            return;
+        ActiveMeasurementsActivity.setEnabledButtons(false);
+        checkServer();
+    }
+
+    private void checkServer() {
+        CheckServerTask checkServerTask = new CheckServerTask(context) {
+            @Override
+            public void handleResponse(int responseCode) {
+                if (responseCode == HttpURLConnection.HTTP_OK)
+                    startSpeedTest();
+                else
+                    selectServer();
+            }
+        };
+        checkServerTask.execute();
+    }
+
+    public void startSpeedTest() {
+        FragmentManager fm = getActivity().getSupportFragmentManager();
+        FragmentTransaction ft = fm.beginTransaction();
+        Fragment prev = fm.findFragmentByTag("speedTestDialog");
+        if (prev != null) {
+            ft.remove(prev);
+        }
+        ft.addToBackStack(null);
+
+        SpeedTestDialog newFragment = new SpeedTestDialog();
+        newFragment.show(ft, "speedTestDialog");
+    }
+
+    private void selectServer(){
+        ActiveServersTask activeServersTask = new ActiveServersTask(context) {
+            @Override
+            public void handleActiveServers(Bundle bundle) {
+                bundle.putBoolean("shouldExecute", true);
+                FragmentManager fm = getActivity().getSupportFragmentManager();
+                ActiveServersDialog dialog = new ActiveServersDialog();
+                dialog.setArguments(bundle);
+                dialog.setCancelable(false);
+                dialog.show(fm, null);
+            }
+        };
+        activeServersTask.execute();
     }
 }
