+++ app/src/main/res/layout/layout_connectivity_test_delete.xml
@@ -1,7 +1,8 @@
 <ImageView xmlns:android="http://schemas.android.com/apk/res/android"
     android:gravity="center"
     android:id="@+id/delete_image"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:src="@drawable/ic_delete_black"
+    android:layout_width="20dp"
+    android:layout_height="20dp"
+    android:src="@drawable/ic_clear_black"
+    android:alpha="0.5"
     android:layout_gravity="center" />
